+++ crates/fuel-core/src/database/block.rs
@@ -3,7 +3,7 @@ use crate::{
         database_description::off_chain::OffChain,
         Database,
     },
-    fuel_core_graphql_api::storage::blocks::FuelBlockSecondaryKeyBlockHeights,
+    fuel_core_graphql_api::storage::blocks::FuelBlockIdsToHeights,
 };
 use fuel_core_storage::{
     iter::IterDirection,
@@ -38,7 +38,7 @@ use std::borrow::Cow;
 
 impl Database<OffChain> {
     pub fn get_block_height(&self, id: &BlockId) -> StorageResult<Option<BlockHeight>> {
-        self.storage::<FuelBlockSecondaryKeyBlockHeights>()
+        self.storage::<FuelBlockIdsToHeights>()
             .get(id)
             .map(|v| v.map(|v| v.into_owned()))
     }