+++ crates/fuel-core/src/query/block.rs
@@ -1,7 +1,4 @@
-use crate::{
-    fuel_core_graphql_api::ports::OnChainDatabase,
-    graphql_api::ports::OffChainDatabase,
-};
+use crate::fuel_core_graphql_api::ports::OnChainDatabase;
 use fuel_core_storage::{
     iter::{
         BoxedIter,
@@ -27,7 +24,7 @@ pub trait SimpleBlockData: Send + Sync {
     fn block(&self, id: &BlockHeight) -> StorageResult<CompressedBlock>;
 }
 
-impl<D: OnChainDatabase + OffChainDatabase + ?Sized> SimpleBlockData for D {
+impl<D: OnChainDatabase + ?Sized> SimpleBlockData for D {
     fn block(&self, id: &BlockHeight) -> StorageResult<CompressedBlock> {
         let block = self
             .storage::<FuelBlocks>()
@@ -53,7 +50,7 @@ pub trait BlockQueryData: Send + Sync + SimpleBlockData {
     fn consensus(&self, id: &BlockHeight) -> StorageResult<Consensus>;
 }
 
-impl<D: OnChainDatabase + OffChainDatabase + ?Sized> BlockQueryData for D {
+impl<D: OnChainDatabase + ?Sized> BlockQueryData for D {
     fn latest_block_height(&self) -> StorageResult<BlockHeight> {
         self.latest_height()
     }